Introduce argument file in pipeline
